UC#1 Create documentation

+++ src/RestCountriesFacade/Endpoints.cs
@@ -9,7 +9,7 @@ namespace RestCountriesFacade
 		{
 			app.MapGet("/retrieve", async ( 
 				string? filter, 
-				int? populationMillions, 
+				decimal? populationMillions, 
 				string? sort, 
 				int? take,
 				RestCountriesClient client) =>
